third_party/blink/web_tests/platform/generic/fast/html/unknown-tag-expected.txt

Test HTMLUnknownElement

This test verifies the following:

The HTMLUnknownElement interface is used for HTML elements that are not defined by the HTML5 specification (or other applicable specifications).
The HTMLUnknownElement interface is not used for HTML elements that are defined by the HTML5 specification (or other applicable specifications).

PASS DynamicElements["div"] instanceof HTMLElement is true
PASS DynamicElements["div"] instanceof HTMLUnknownElement is false
PASS DynamicElements["font"] instanceof HTMLElement is true
PASS DynamicElements["font"] instanceof HTMLUnknownElement is false
PASS DynamicElements["h1"] instanceof HTMLElement is true
PASS DynamicElements["h1"] instanceof HTMLUnknownElement is false
PASS DynamicElements["table"] instanceof HTMLElement is true
PASS DynamicElements["table"] instanceof HTMLUnknownElement is false
PASS DynamicElements["foo1"] instanceof HTMLElement is true
PASS DynamicElements["foo1"] instanceof HTMLUnknownElement is true
PASS DynamicElements["foo2"] instanceof HTMLElement is true
PASS DynamicElements["foo2"] instanceof HTMLUnknownElement is true
PASS staticElements["div"] instanceof HTMLUnknownElement is false
PASS staticElements["font"] instanceof HTMLUnknownElement is false
PASS staticElements["h1"] instanceof HTMLUnknownElement is false
PASS staticElements["table"] instanceof HTMLUnknownElement is false
PASS staticElements["foo1"] instanceof HTMLUnknownElement is true
PASS staticElements["foo2"] instanceof HTMLUnknownElement is true
PASS successfullyParsed is true

TEST COMPLETE
